pyproject.toml
[project]
name = "thosono-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "google-genai>=1.69.0",
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "uvicorn>=0.42.0",
]
